LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-redirect-credentials.any-expected.txt

CONSOLE MESSAGE: line 21: ReferenceError: Can't find variable: dirname

Harness Error (FAIL), message = ReferenceError: Can't find variable: dirname
